requirements.txt
Django==5.2.4
djangorestframework==3.16.0
pillow==11.3.0
PyYAML==6.0.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
user-agents==2.2.0